requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
urllib3==1.26.2
Werkzeug==1.0.1
